refactor: export all data related functions to sep file
